# docs/swagger/swagger.yaml
basePath: /api/v3
definitions:
  user-service_internal_model.CreateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  user-service_internal_model.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: Invalid parameters
        type: string
      message:
        example: Bad Request
        type: string
    type: object
  user-service_internal_model.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  user-service_internal_model.LoginResponse:
    properties:
      expires_in:
        example: 3600
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/user-service_internal_model.User'
    type: object
  user-service_internal_model.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
  user-service_internal_model.UpdateUserRequest:
    properties:
      avatar:
        example: https://example.com/avatar.png
        type: string
      email:
        example: john@example.com
        type: string
      phone:
        example: "13800138000"
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    type: object
  user-service_internal_model.User:
    properties:
      avatar:
        example: https://example.com/avatar.png
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      phone:
        example: "13800138000"
        type: string
      status:
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  user-service_internal_model.UserListResponse:
    properties:
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      users:
        items:
          $ref: '#/definitions/user-service_internal_model.User'
        type: array
    type: object
host: localhost:8001
info:
  contact:
    email: support@company.com
    name: API Support
    url: http://www.company.com/support
  description: 用户管理服务 - 提供用户注册、登录、信息管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User Service API
  version: 3.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录
      parameters:
      - description: 登录凭证
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user-service_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-service_internal_model.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      summary: 用户登录
      tags:
      - Auth
  /health:
    get:
      description: 检查服务是否正常运行
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 健康检查
      tags:
      - System
  /users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-service_internal_model.UserListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user-service_internal_model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/user-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-service_internal_model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-service_internal_model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-service_internal_model.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 根据用户ID更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user-service_internal_model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-service_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-service_internal_model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"